fixed a million tests
