+++ tests/server/blueprints/institutes/test_institutes_views.py
@@ -38,7 +38,9 @@ def test_institute(app, user_obj, institute_obj):
         client.get(url_for("auto_login"))
 
         # WHEN accessing the cases page (GET method)
-        resp = client.get(url_for("overview.institute", institute_id=institute_obj["internal_id"]))
+        resp = client.get(
+            url_for("overview.institute_settings", institute_id=institute_obj["internal_id"])
+        )
 
         # THEN it should return a page
         assert resp.status_code == 200
@@ -59,7 +61,7 @@ def test_institute(app, user_obj, institute_obj):
 
         # via POST request
         resp = client.post(
-            url_for("overview.institute", institute_id=institute_obj["internal_id"]),
+            url_for("overview.institute_settings", institute_id=institute_obj["internal_id"]),
             data=form_data,
         )
         assert resp.status_code == 200
@@ -73,6 +75,166 @@ def test_institute(app, user_obj, institute_obj):
         assert updated_institute["cohorts"] == form_data["cohorts"]
         assert updated_institute["collaborators"] == form_data["institutes"]
         assert len(updated_institute["phenotype_groups"]) == 2  # one for each HPO term
+
+
+def test_cases(app, institute_obj):
+    # GIVEN an initialized app
+    # GIVEN a valid user and institute
+
+    with app.test_client() as client:
+        # GIVEN that the user could be logged in
+        resp = client.get(url_for("auto_login"))
+        assert resp.status_code == 200
+
+        # WHEN accessing the cases page
+        resp = client.get(url_for("overview.cases", institute_id=institute_obj["internal_id"]))
+
+        # THEN it should return a page
+        assert resp.status_code == 200
+
+        # test query passing parameters in seach form
+        request_data = {
+            "limit": "100",
+            "skip_assigned": "on",
+            "is_research": "on",
+            "query": "case_id",
+        }
+        resp = client.get(
+            url_for(
+                "overview.cases", institute_id=institute_obj["internal_id"], params=request_data,
+            )
+        )
+        # response should return a page
+        assert resp.status_code == 200
+
+        sorting_options = ["analysis_date", "track", "status"]
+        for option in sorting_options:
+            # test query passing the sorting option to the cases view
+            request_data = {"sort": option}
+            resp = client.get(
+                url_for(
+                    "overview.cases",
+                    institute_id=institute_obj["internal_id"],
+                    params=request_data,
+                )
+            )
+            # response should return a page
+            assert resp.status_code == 200
+
+
+def test_cases_query(app, case_obj, institute_obj):
+    # GIVEN an initialized app
+    # GIVEN a valid user and institute
+
+    slice_query = case_obj["display_name"]
+
+    with app.test_client() as client:
+        # GIVEN that the user could be logged in
+        resp = client.get(url_for("auto_login"))
+        assert resp.status_code == 200
+
+        # WHEN accessing the cases page with a query
+        resp = client.get(
+            url_for("overview.cases", query=slice_query, institute_id=institute_obj["internal_id"],)
+        )
+
+        # THEN it should return a page
+        assert resp.status_code == 200
+
+
+def test_cases_panel_query(app, case_obj, parsed_panel, institute_obj):
+    # GIVEN an initialized app
+    # GIVEN a valid user and institute
+
+    slice_query = parsed_panel["panel_id"]
+
+    with app.test_client() as client:
+        # GIVEN that the user could be logged in
+        resp = client.get(url_for("auto_login"))
+        assert resp.status_code == 200
+
+        # WHEN accessing the cases page with a query
+        resp = client.get(
+            url_for("overview.cases", query=slice_query, institute_id=institute_obj["internal_id"],)
+        )
+
+        # THEN it should return a page
+        assert resp.status_code == 200
+
+
+def test_causatives(app, user_obj, institute_obj, case_obj):
+    # GIVEN an initialized app
+    # GIVEN a valid user and institute
+    # There should be no causative variants for test case:
+    assert "causatives" not in case_obj
+    var1_id = "4c7d5c70d955875504db72ef8e1abe77"  # in POT1 gene
+    var2_id = "e24b65bf27feacec6a81c8e9e19bd5f1"  # in TBX1 gene
+    var_ids = [var1_id, var2_id]
+
+    # for each variant
+    for var_id in var_ids:
+        # update case by marking variant as causative:
+        variant_obj = store.variant(document_id=var_id)
+        store.mark_causative(
+            institute=institute_obj,
+            case=case_obj,
+            user=user_obj,
+            link="causative_var_link/{}".format(variant_obj["_id"]),
+            variant=variant_obj,
+        )
+    updated_case = store.case_collection.find_one({"_id": case_obj["_id"]})
+    # The above variants should be registered as causatives in case object
+    assert updated_case["causatives"] == var_ids
+
+    # Call scout causatives view and check if the above causatives are displayed
+    with app.test_client() as client:
+
+        # GIVEN that the user could be logged in
+        resp = client.get(url_for("auto_login"))
+        assert resp.status_code == 200
+
+        # WHEN accessing the case page
+        resp = client.get(url_for("overview.causatives", institute_id=institute_obj["internal_id"]))
+
+        # THEN it should return a page
+        assert resp.status_code == 200
+        # with variant 1
+        assert var1_id in str(resp.data)
+        # and variant 2
+        assert var2_id in str(resp.data)
+
+        # Filter causatives by gene (POT1)
+        resp = client.get(
+            url_for(
+                "overview.causatives",
+                institute_id=institute_obj["internal_id"],
+                query="17284 | POT1 (DKFZp586D211, hPot1, POT1)",
+            )
+        )
+        # THEN it should return a page
+        assert resp.status_code == 200
+        # with variant 1
+        assert var1_id in str(resp.data)
+        # but NOT variant 2
+        assert var2_id not in str(resp.data)
+
+
+def test_search_institute_variants(app, user_obj, institute_obj):
+    # GIVEN an initialized app
+    # GIVEN a valid user and institute
+
+    with app.test_client() as client:
+        # GIVEN that the user could be logged in
+        resp = client.get(url_for("auto_login"))
+        assert resp.status_code == 200
+
+        # WHEN accessing the dashboard page
+        resp = client.get(
+            url_for("overview.gene_variants", institute_id=institute_obj["internal_id"])
+        )
+
+        # THEN it should return a page
+        assert resp.status_code == 200
 
 
 def test_clinvar_submissions(app, institute_obj, clinvar_variant, clinvar_casedata):
